Uttrekk fra Dynarev til R

dynarev_uttrekk er en funksjon som henter inn data fra Dynarev til R så lenge man oppgir delregisternummer og skjemanavn. Det er kun enheter som er satt som aktive (AKTIV = 1) som blir lastet ned. Det er mulig å hente skjemadata (altså variablene i selve skjemaet) og SFU-data (enhetsinformasjon). Man kan også gjennomføre dublettsjekker og velge mellom reviderte data og rådata.

Usage

dynarev_uttrekk(
  delregnr,
  skjema,
  enhets_type = c("FRTK", "BEDR"),
  skjema_cols = T,
  skjema_sfu_merge = F,
  dublettsjekk = F,
  sfu_cols = F,
  con_ask = T,
  raadata = F
)

Arguments

delregnr

Numerisk vektor med delregisternummer.

skjema

Karakter vektor med skjemanavn.

enhets_type

Karakter vektor med enhetstype.

skjema_cols

TRUE = alle kolonner fra skjema. Karakter vektor for kun utvalgte kolonner.

skjema_sfu_merge

TRUE = skjemadata og SFU merges sammen til én data.frame.

dublettsjekk

TRUE = dublettsjekk etter ENHETS_ID. Karakter vektor for dublettsjekk etter valgte kolonner.

sfu_cols

TRUE = alle kolonner fra SFU. Karakter vektor for kun utvalgte kolonner.

con_ask

TRUE = spør om passord for å koble til DB1P og laster inn data. Dersom con_ask = "con" kobles det til DB1P uten å laste inn data.

raadata

FALSE returnerer reviderte data, TRUE returnerer rådata.

Value

data.frame med valgt data fra Dynarev, dersom enten skjema_cols eller sfu_cols = TRUE. Om både skjema_cols og sfu_cols = TRUE returneres en liste med to datasett [1] skjema_cols og [2] sfu_cols. Om skjema_cols, sfu_cols og skjema_sfu_merge returneres det en data frame.

Examples

if (FALSE) {
dynarev <- dynarev_uttrekk(delregnr = 2421,
                          skjema = "HELSE41",
                          skjema_cols = T)
}