Hent metadata for én variabel fra Datadoc

Leser en DataDoc-fil og henter det fullstendige metadataobjektet for en bestemt variabel.

Usage

metadata_variable(filsti, variabel)

Arguments

filsti

En tekststreng med filstien til Parquet-filen. Filstien til den tilhørende DataDoc-filen utledes med datadoc_path().

variabel

En tekststreng med kortnavnet til variabelen som metadata skal hentes for.

Value

En liste med metadataene som er registrert for variabelen i DataDoc-filen.

Details

Funksjonen leser DataDoc-filen med jsonlite::fromJSON() og søker etter en variabel der feltet short_name er identisk med verdien i variabel.

Funksjonen stopper med en feilmelding dersom variabelen ikke finnes, eller dersom flere variabler har samme kortnavn.

See also

Examples

if (FALSE) { # \dontrun{
metadata <- metadata_variable(
  filsti = "data/personell.parquet",
  variabel = "kjoenn"
)

metadata$name
metadata$classification_uri
} # }